+++ src/Ext/TechnoType/Body.h
@@ -65,8 +65,6 @@ public:
 		Nullable<WeaponTypeClass*> WarpInMinRangeWeapon;
 		Nullable<WeaponTypeClass*> WarpOutWeapon;
 		Valueable<bool> WarpInWeapon_UseDistanceAsDamage;
-		Valueable<bool> WarpInWeapon_FireAsSelf;
-		Valueable<bool> WarpOutWeapon_FireAsSelf;
 
 		ValueableVector<AnimTypeClass*> OreGathering_Anims;
 		ValueableVector<int> OreGathering_Tiberiums;
@@ -158,8 +156,6 @@ public:
 			, WarpInMinRangeWeapon {}
 			, WarpOutWeapon {}
 			, WarpInWeapon_UseDistanceAsDamage { false }
-			, WarpInWeapon_FireAsSelf { true }
-			, WarpOutWeapon_FireAsSelf { true }
 			, OreGathering_Anims {}
 			, OreGathering_Tiberiums {}
 			, OreGathering_FramesPerDir {}